Fix markdown typo and use consistent abbrev. style
Typo on line 61 caused some Markdown to not be rendered.

Abbreviations for academic degrees on lines 61 and 77 were inconsistent; the former used both punctuation and spacing and the latter used neither. Of the style references I found, most (including the Chicago Manual of Style and the University of Oxford) preferred PhD to Ph.D. and none allowed spaces within such abbreviations.