date change for testing; thanked balpha
Even dev.blog was respecting the 12/14 date, despite the fact that draft was set to true, so I'm changing the date to yesterday just to make sure the post shows up on the blog without any unpleasant formatting surprises. Must remember to change this back when setting draft to false tomorrow night...

Also fixed the oversight of not giving balpha any props in the original draft.