added link to companion meta post
Finally finished writing the companion post on meta, and added a link at the bottom of this blog post as promised in the original text. Also changed the "description" metadata (because apparently that's what shows up in the preview area on the blog front page) and fixed a few minor typos.