Fix bad escape of MD-emphasized asterisk
The desired result of `*\* foo*` is to apply MD emphasis to a footnote. The parser apparently gives higher priority to closing emphasis delimiters than to escapes, because the actual result is an italic backslash instead of an escaped asterisk. This PR achieves the desired result by applying MD emphasis only the text of the footnote and not the symbol.